add_odometryを追加
・コンパイルは通った
・add_chassisとの整合性は未確認

+++ program/whole/inc/ext_var.hpp
@@ -8,8 +8,8 @@ extern jibiki::ShareVar<short> g_rev_tgt[4];       /* 回転数目標値 */
 extern jibiki::ShareVar<short> g_rev_curr[4];      /* 回転数現在地 */
 extern jibiki::ShareVar<int16_t> g_rot_tgt[4];       /* 回転角目標値 */
 extern jibiki::ShareVar<int16_t> g_rot_curr[4];      /* 回転角現在地 */
-extern jibiki::ShareVar<int64_t>g_dist_tgt[2];     /* 移動距離目標値 */
-extern jibiki::ShareVar<int64_t>g_dist_curr[2];    /* 移動距離 */
+extern jibiki::ShareVar<int32_t>g_dist_tgt[2];     /* 移動距離目標値 */
+extern jibiki::ShareVar<int32_t>g_dist_curr[2];    /* 移動距離 */
 extern jibiki::ShareVar<bool>  g_odometry_flag[2]; /* odometry_リセットフラグ */
 extern jibiki::ShareVar<bool>g_limit[8];           /* limitスイッチのon/off */
 extern SwitchData g_switch_data;                   /* スイッチスレーブのデータ */
